Switch OSX to use std::filesystem::space() instead of statvfs()
Now we have all platforms sharing a single block of code for checking
disk space, woo!

+++ VS/Files/VS_File.h
@@ -125,9 +125,9 @@ public:
 
 	struct DiskStats
 	{
-		size_t freeBytes;      // bytes without things stored in them.
-		size_t availableBytes; // bytes which *we* can write into.  (may be lower than 'freeBytes'!)
-		size_t capacityBytes;  // total bytes on the disk.
+		uintmax_t freeBytes;      // bytes without things stored in them.
+		uintmax_t availableBytes; // bytes which *we* can write into.  (may be lower than 'freeBytes'!)
+		uintmax_t capacityBytes;  // total bytes on the disk.
 	};
 
 	static void LogDiskStats(); // logs information about our disk.
